+++ SeriesSyndex/pmse.py
@@ -113,8 +113,7 @@ class pMSEEvaluator:
             losses = []
             for batch in train_data_loader:
                 (static_vars, series_vars), labels = batch
-                
-                outputs = self.model(series_vars)
+                outputs = self.model(series_vars.float())
 
                 loss = loss_fn(outputs.squeeze(), labels.float())
                 losses.append(loss.item())
@@ -139,7 +138,7 @@ class pMSEEvaluator:
         for batch in test_data_loader:
             (static_vars, series_vars), labels = batch
                 
-            outputs = self.model(series_vars)
+            outputs = self.model(series_vars.float())
             loss = loss_fn(outputs.squeeze(), labels.float())
             losses.append(loss.item())
             target_labels.append(labels.cpu())
